pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "srdownscalling"
version = "0.1.0"
description = "Super-Resolution for WRF Wind Fields using ESGAN + Attention"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "oriolIA", email = "oriolbot2@gmail.com"}
]
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "numpy>=1.24.0",
    "netCDF4>=1.6.0",
    "xarray>=2023.0.0",
    "pandas>=2.0.0",
    "scikit-image>=0.21.0",
    "scipy>=1.11.0",
    "matplotlib>=3.7.0",
    "tqdm>=4.65.0",
    "pyyaml>=6.0",
    "hydra-core>=1.3.0",
    "wandb>=0.15.0",
    "albumentations>=1.3.0",
    "einops>=0.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
